fix: Fix "Returning Any from functions" errors in backend and core modules (186 → 173 errors)
- queue_service.py: Cast get_queue_info() return to Dict[str, Any]
- playback_service.py: Cast model_dump() return to Dict[str, Any]
- analysis_extractor.py: Cast cache.get() and cache.get_stats() returns (3 fixes)
- routers/dependencies.py: Cast return values to LibraryManager and EnhancedAudioPlayer
- routers/library.py: Cast factory.stats.get_library_stats() return
- routers/playlists.py: Cast playlist.to_dict() return
- config/middleware.py: Cast call_next() response to Response
- continuous_space.py: Wrap float returns with float() to narrow numpy type (3 fixes)

Pattern: Using cast() for dynamically accessed return values (hasattr checks, method calls on Any types) and float() for numpy scalar returns. Reduced mypy errors from 186 to 173 (13 errors fixed).

+++ auralis/dsp/utils/spectral.py
@@ -59,7 +59,7 @@ def spectral_centroid(audio: np.ndarray, sample_rate: int = 44100) -> float:
     else:
         centroid = sample_rate // 4  # Default to quarter of sample rate
 
-    return centroid
+    return float(centroid)
 
 
 def spectral_rolloff(audio: np.ndarray,
@@ -96,9 +96,9 @@ def spectral_rolloff(audio: np.ndarray,
         rolloff_threshold = rolloff_percent * total_energy
         rolloff_idx = np.where(cumulative_energy >= rolloff_threshold)[0]
         if len(rolloff_idx) > 0:
-            return freqs[rolloff_idx[0]]
+            return float(freqs[rolloff_idx[0]])
 
-    return sample_rate // 2  # Default to Nyquist frequency
+    return float(sample_rate // 2)  # Default to Nyquist frequency
 
 
 def zero_crossing_rate(audio: np.ndarray) -> float:
@@ -119,7 +119,7 @@ def zero_crossing_rate(audio: np.ndarray) -> float:
 
     # Count zero crossings
     zero_crossings = np.sum(np.diff(np.sign(audio)) != 0)
-    return zero_crossings / (2 * len(audio))
+    return float(zero_crossings / (2 * len(audio)))
 
 
 def crest_factor(audio: np.ndarray) -> float:
